moved examples from the __init__ to an example folder
